requirements.txt
Requests>=2.19.0
simpletransformers==0.18.3
pandas==0.25.3
tqdm==4.41.1
spacy==2.2.3
python-dotenv==0.10.4
O365==2.0.6
IMAPClient==2.1.0